Major update to all games
Added levelling! Now you get more cash daily and passively the more you play!
If you bet 1 coin on a game, you only get 1% of the normal xp gain, but if you bet 100 or more, you gain the normal amout of xp
Added doubling down in blackjack, and a button for it
Adjusted daily gains
Added debt! For now its owners only, but later everyone will be able to get a loan from the bank, which then they will have to pay off, either with $paydebt, or by the bank taking 10% of each win.